students/K33402/Prokhorov_Nikolay/LR3/superweather/requirements.txt
Django~=3.2.11
django-cors-headers
django-extensions
django-filter~=2.4.0
djangorestframework~=3.13.1
djangorestframework-camel-case
Werkzeug
drf-yasg
djoser
djangorestframework_simplejwt
requests~=2.24.0